requirements.txt
pandas>=2.1
openpyxl>=3.1
python-dateutil>=2.9